common: Reformat some comments in iobuf.c
--

Comments with stars on the left side are easier to read.  Also a
long comment in the form

  if (...)
     /* This is a comment
      * for the next block */
    {
       ...
    }

makes it hard to see the start of a block or hides a forgotten block.
In contrast to GNU common use we don't want this but put the comment
either into the block or with adjusted wording above the condition.