scd:nks: Make newer TCOS signature cards work.
* scd/app-nks.c: Make appversion 232 an alias for 15.
--

I received a new QES signature card and it turned out that the
appversion was changed from 15 to 232.  I have no further information
but simply tried this patch and it worked as far as I can check: I was
able to init the PINs and PUKs using gpg-card and could also install
the extra certificates using the writcert command.  Signing with this
card also works.

I have asked Telesec for updated specs.