common:dotlock: Escalate a warning message up to INFO from DEBUG.
* common/dotlock.c (dotlock_take): Use my_info_1 when already locked.
(dotlock_release): Likewise.

--

Since recursive lock is not intended with dotlock (yet), it's a
programming error when detecting already locked state on the call of
dotlock_take.  Also, it's a programming error when detecting
non-locked state on the call of dotlock_release.  Use of my_debug_1 is
not relevant here.

Signed-off-by: NIIBE Yutaka <gniibe@fsij.org>